requirements.txt
streamlit==1.30.0
supabase==2.7.4
python-dotenv==1.0.0
pillow==10.2.0
plotly==5.18.0
pandas==2.1.4
bcrypt==4.1.2
pillow-heif==0.15.0
